Separate out mod solving from discovery.
