+++ src/main/java/org/quiltmc/loader/impl/solver/LoadOption.java
@@ -1,0 +1,7 @@
+package org.quiltmc.loader.impl.solver;
+
+/** Base definition of something that can either be completely loaded or not loaded. (Usually this is just a mod jar
+ * file, but in the future this might refer to something else that loader has control over). */
+abstract class LoadOption {
+
+}
